INDY-2098: Yet another attempt to fix failing tests
